perf: Shrink Permanent.Counters from [N]int to [N]uint8
The change is a memory optimization — int is 8 bytes on 64-bit, uint8 is 1 byte, so the
counters array drops from ~240 bytes to ~30 bytes per permanent. Counter values in MTG rarely
exceed 255, so uint8 is sufficient. All call sites that need int (comparisons, arithmetic, API
boundaries) get explicit int() casts.

+++ cards/antiquities/creatures.go
@@ -829,7 +829,7 @@ func registerCreatures() {
 						if src == nil {
 							return nil
 						}
-						maxCounters := src.Counters[P1P1]
+						maxCounters := int(src.Counters[P1P1])
 						if maxCounters == 0 {
 							return nil
 						}
@@ -838,7 +838,7 @@ func registerCreatures() {
 						if n == 0 {
 							return nil
 						}
-						src.Counters[P1P1] -= n
+						src.Counters[P1P1] -= uint8(n)
 						for range n {
 							token := NewToken("Tetravite", 1, 1,
 								[]CardType{TypeArtifact, TypeCreature},
@@ -880,7 +880,7 @@ func registerCreatures() {
 								g.ExilePermanent(tetravites[i])
 							}
 						}
-						src.Counters[P1P1] += n
+						src.Counters[P1P1] += uint8(n)
 						return nil
 					}), true,
 			)),
